+++ src/revisionGraph/flow/index.ts
@@ -6,6 +6,7 @@ export * from './flowDiagnostics';
 export * from './flowEqualization';
 export * from './flowPromotionChecks';
 export * from './flowPullRequestContext';
+export * from './flowReleaseBranch';
 export * from './flowState';
 export * from './flowTransitionPolicy';
 export * from './flowTypes';